+++ test/WhatsUpToday.Core.Data.Test/Common/StudentContext.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore;
+
+namespace WhatsUpToday.Core.Data.Test.Common;
+
+public class StudentContext : DbContext
+{
+    public StudentContext()
+    { }
+
+    public StudentContext(DbContextOptions<StudentContext> options) : base(options)
+    { }
+
+    public DbSet<Student> Students { get; set; }
+}