+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 
+**1.0.3**
+* Core:
+  * Add exception helper class Ensure.
+  * Add string extensions.
+* Data:
+  * Add entity interfaces for AutoSave (see IAutoSaveEntity)
+  * Add DbContext extensions for AutoSave, which automatically sets the Entity creation/modification dates and user for tracked entities based on AutoSave interfaces.
+* Test:
+  * Add test libraries for Core and Data.
+  * Add unit tests for AutoSave extension.
+ 
 **1.0.2**
 * Core: Adds generic Entity types.
 * Data: Adds generic services for Entity types.